Separate upgrade code from meta settings
