+++ collect_app/src/main/java/org/odk/collect/android/application/initialization/upgrade/AppUpgrader.kt
@@ -1,20 +1,17 @@
 package org.odk.collect.android.application.initialization.upgrade
 
 import org.odk.collect.android.BuildConfig
-import org.odk.collect.android.utilities.InstallDetector
-import org.odk.collect.android.utilities.LaunchState
-import org.odk.collect.android.utilities.VersionCodeLaunchState
 import org.odk.collect.shared.Settings
 
 class AppUpgrader(
-    private val metaSettings: Settings,
+    private val settings: Settings,
     private val launchState: LaunchState,
     private val upgrades: List<Upgrade>
 ) {
 
-    constructor(metaSettings: Settings, installDetector: InstallDetector, upgrades: List<Upgrade>) : this(
-        metaSettings,
-        VersionCodeLaunchState(metaSettings, BuildConfig.VERSION_CODE, installDetector),
+    constructor(key: String, settings: Settings, installDetector: InstallDetector, upgrades: List<Upgrade>) : this(
+        settings,
+        VersionCodeLaunchState(key, settings, BuildConfig.VERSION_CODE, installDetector),
         upgrades
     )
 
@@ -25,15 +22,15 @@ class AppUpgrader(
 
                 if (key == null) {
                     it.run()
-                } else if (!metaSettings.getBoolean(key)) {
+                } else if (!settings.getBoolean(key)) {
                     it.run()
-                    metaSettings.save(key, true)
+                    settings.save(key, true)
                 }
             }
         } else {
             upgrades.forEach {
                 it.key()?.let { key ->
-                    metaSettings.save(key, true)
+                    settings.save(key, true)
                 }
             }
         }
