+++ collect_app/src/test/java/org/odk/collect/android/application/initialization/upgrade/AppUpgraderTest.kt
@@ -6,7 +6,6 @@ import org.mockito.kotlin.mock
 import org.mockito.kotlin.times
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
-import org.odk.collect.android.utilities.LaunchState
 import org.odk.collect.testshared.InMemSettings
 
 class AppUpgraderTest {
