+++ collect_app/src/test/java/org/odk/collect/android/application/initialization/upgrade/VersionCodeLaunchStateTest.kt
@@ -5,49 +5,48 @@ import org.hamcrest.Matchers.equalTo
 import org.junit.Test
 import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.mock
-import org.odk.collect.android.preferences.keys.MetaKeys
-import org.odk.collect.android.utilities.VersionCodeLaunchState
 import org.odk.collect.testshared.InMemSettings
 
 class VersionCodeLaunchStateTest {
 
     @Test
-    fun `isUpgradedFirstLaunch() returns false for empty meta settings`() {
-        val appStateProvider = VersionCodeLaunchState(InMemSettings(), 1)
+    fun `isUpgradedFirstLaunch() returns false for empty settings`() {
+        val appStateProvider = VersionCodeLaunchState("key", InMemSettings(), 1)
         assertThat(appStateProvider.isUpgradedFirstLaunch(), equalTo(false))
     }
 
     @Test
     fun `isUpgradedFirstLaunch() returns false if last launched is equal to current version`() {
         val inMemSettings = InMemSettings()
-        inMemSettings.save(MetaKeys.LAST_LAUNCHED, 1)
+        inMemSettings.save("key", 1)
 
-        val appStateProvider = VersionCodeLaunchState(inMemSettings, 1)
+        val appStateProvider = VersionCodeLaunchState("key", inMemSettings, 1)
         assertThat(appStateProvider.isUpgradedFirstLaunch(), equalTo(false))
     }
 
     @Test
     fun `isUpgradedFirstLaunch() returns true if last launched is less than current version`() {
         val inMemSettings = InMemSettings()
-        inMemSettings.save(MetaKeys.LAST_LAUNCHED, 1)
+        inMemSettings.save("key", 1)
 
-        val appStateProvider = VersionCodeLaunchState(inMemSettings, 2)
+        val appStateProvider = VersionCodeLaunchState("key", inMemSettings, 2)
         assertThat(appStateProvider.isUpgradedFirstLaunch(), equalTo(true))
     }
 
     @Test
     fun `isUpgradedFirstLaunch() returns false if last launched is less than current version after appLaunched()`() {
         val inMemSettings = InMemSettings()
-        inMemSettings.save(MetaKeys.LAST_LAUNCHED, 1)
+        inMemSettings.save("key", 1)
 
-        val appStateProvider = VersionCodeLaunchState(inMemSettings, 2)
+        val appStateProvider = VersionCodeLaunchState("key", inMemSettings, 2)
         appStateProvider.appLaunched()
         assertThat(appStateProvider.isUpgradedFirstLaunch(), equalTo(false))
     }
 
     @Test
-    fun `isUpgradedFirstLaunch() returns true for empty meta settings when legacy install detected`() {
+    fun `isUpgradedFirstLaunch() returns true for empty settings when legacy install detected`() {
         val appStateProvider = VersionCodeLaunchState(
+            "key",
             InMemSettings(),
             1,
             mock { on { installDetected() } doReturn true }
@@ -57,8 +56,9 @@ class VersionCodeLaunchStateTest {
     }
 
     @Test
-    fun `isUpgradedFirstLaunch() returns false for empty meta settings when legacy install not detected`() {
+    fun `isUpgradedFirstLaunch() returns false for empty settings when legacy install not detected`() {
         val appStateProvider = VersionCodeLaunchState(
+            "key",
             InMemSettings(),
             1,
             mock { on { installDetected() } doReturn false }
